requirements.txt
groq==0.11.0
pandas==2.2.0
selenium==4.18.1
python-dotenv==1.0.1
openpyxl==3.1.2
